Refactor: Split things into separate files.
