+++ standard_training/grading/__init__.py
@@ -3,7 +3,9 @@ This module exists to provide common types of gading for exercises
 als allow composition of them.
 """
 from .event_detector import PlayerEventDetector, PlayerEvent, PlayerEventType
-from .grader import Grader, GraderExercise, TrainingTickPacket
+from .training_tick_packet import TrainingTickPacket
+from .grader import Grader
+from .grader_exercise import GraderExercise
 from .compound_grader import CompoundGrader
 from .timeout import FailOnTimeout, PassOnTimeout
 from .goal_grader import StrikerGrader, GoalieGrader
